# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthSessionResponseDto": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "refreshTokenExpiresAt",
          "userId"
        ],
        "type": "object"
      },
...
      "ForgotPasswordDto": {
        "properties": {
          "email": {
            "maxLength": 320,
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "MessageResponseDto": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "RefreshDto": {
        "properties": {
          "deviceLabel": {
            "maxLength": 120,
            "type": "string"
          },
          "refreshToken": {
            "maxLength": 512,
            "minLength": 20,
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "ResetPasswordDto": {
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 12,
            "type": "string"
          },
          "token": {
            "maxLength": 512,
            "minLength": 20,
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "HTTP API for this NestJS service",
    "title": "Service API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "Auth.forgotPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            },
            "description": "Acknowledged"
          }
        },
        "security": [
          {}
        ],
        "summary": "Request a password reset (generic response)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/logout-all": {
      "post": {
        "operationId": "Auth.logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            },
            "description": "Sessions revoked"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Revoke all sessions for the current user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "Auth.refresh",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseDto"
                }
              }
            },
            "description": "Session rotated"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "security": [
          {}
        ],
        "summary": "Rotate a refresh session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "Auth.resetPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            },
            "description": "Password reset"
          }
        },
        "security": [
          {}
        ],
        "summary": "Reset a password with a one-time token",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "servers": [],
...
}